+++ vertx-core/src/main/java/io/vertx/core/eventbus/impl/clustered/ClusteredEventBus.java
@@ -24,8 +24,8 @@ import io.vertx.core.internal.logging.Logger;
 import io.vertx.core.internal.logging.LoggerFactory;
 import io.vertx.core.internal.net.NetSocketInternal;
 import io.vertx.core.net.*;
-import io.vertx.core.net.impl.NetClientBuilder;
-import io.vertx.core.net.impl.NetServerInternal;
+import io.vertx.core.net.impl.tcp.NetClientBuilder;
+import io.vertx.core.net.impl.tcp.NetServerInternal;
 import io.vertx.core.spi.cluster.ClusterManager;
 import io.vertx.core.spi.cluster.NodeInfo;
 import io.vertx.core.spi.cluster.RegistrationInfo;
