+++ vertx-core/src/main/java/io/vertx/core/http/impl/Http1xClientConnection.java
@@ -46,7 +46,7 @@ import io.vertx.core.internal.net.NetSocketInternal;
 import io.vertx.core.net.HostAndPort;
 import io.vertx.core.net.SocketAddress;
 import io.vertx.core.net.impl.MessageWrite;
-import io.vertx.core.net.impl.NetSocketImpl;
+import io.vertx.core.net.impl.tcp.NetSocketImpl;
 import io.vertx.core.net.impl.VertxHandler;
 import io.vertx.core.spi.metrics.ClientMetrics;
 import io.vertx.core.spi.metrics.HttpClientMetrics;
