+++ vertx-core/src/main/java/io/vertx/core/impl/VertxImpl.java
@@ -45,6 +45,10 @@ import io.vertx.core.internal.threadchecker.BlockedThreadChecker;
 import io.vertx.core.net.*;
 import io.vertx.core.net.impl.*;
 import io.vertx.core.impl.transports.NioTransport;
+import io.vertx.core.net.impl.tcp.CleanableNetClient;
+import io.vertx.core.net.impl.tcp.NetClientBuilder;
+import io.vertx.core.net.impl.tcp.NetServerImpl;
+import io.vertx.core.net.impl.tcp.NetServerInternal;
 import io.vertx.core.spi.context.executor.EventExecutorProvider;
 import io.vertx.core.spi.context.storage.AccessMode;
 import io.vertx.core.spi.context.storage.ContextLocal;
