+++ vertx-core/src/main/java/io/vertx/core/net/impl/tcp/ChannelProvider.java
@@ -9,7 +9,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
  */
 
-package io.vertx.core.net.impl;
+package io.vertx.core.net.impl.tcp;
 
 import io.netty.bootstrap.Bootstrap;
 import io.netty.channel.Channel;
